Nexus-mvp 2.0/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
email-validator==2.2.0
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
boto3==1.35.20
httpx==0.27.2

